ai_server/requirements.txt
blinker==1.6.2
boto3==1.26.130
botocore==1.29.130
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
emoji==1.2.0
filelock==3.12.0
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
fsspec==2023.4.0
greenlet==2.0.2
gunicorn==20.1.0
huggingface-hub==0.14.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kss==4.5.1
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
packaging==23.1
pecab==1.0.8
pluggy==1.0.0
pyarrow==12.0.0
PyMySQL==1.0.3
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
s3transfer==0.6.1
sentencepiece==0.1.99
six==1.16.0
SQLAlchemy==2.0.12
sympy==1.11.1
tokenizers==0.13.3
torch==2.0.0
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.25.4
Werkzeug==2.3.3
